LayoutTests/fast/canvas/webgl/array-setters-expected.txt

Test all permutations of WebGLArray setters to make sure values don't get truncated

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Regression test for https://bugs.webkit.org/show_bug.cgi?id=33350 : WebGLArray subclasses do the wrong conversion in indexSetter
Testing WebGLByteArray
PASS webGLArray is array
PASS webGLArray[0] is array[0]
PASS webGLArray.get(0) is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray.get(1) is array[1]
PASS webGLArray[0] is 0
PASS webGLArray.get(0) is 0
PASS webGLArray[1] is 0
PASS webGLArray.get(1) is 0
PASS webGLArray[0] is array[0]
PASS webGLArray.get(0) is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray.get(1) is array[1]
Testing WebGLUnsignedByteArray
PASS webGLArray is array
PASS webGLArray[0] is array[0]
PASS webGLArray.get(0) is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray.get(1) is array[1]
PASS webGLArray[0] is 0
PASS webGLArray.get(0) is 0
PASS webGLArray[1] is 0
PASS webGLArray.get(1) is 0
PASS webGLArray[0] is array[0]
PASS webGLArray.get(0) is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray.get(1) is array[1]
Testing WebGLShortArray
PASS webGLArray is array
PASS webGLArray[0] is array[0]
PASS webGLArray.get(0) is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray.get(1) is array[1]
PASS webGLArray[0] is 0
PASS webGLArray.get(0) is 0
PASS webGLArray[1] is 0
PASS webGLArray.get(1) is 0
PASS webGLArray[0] is array[0]
PASS webGLArray.get(0) is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray.get(1) is array[1]
Testing WebGLUnsignedShortArray
PASS webGLArray is array
PASS webGLArray[0] is array[0]
PASS webGLArray.get(0) is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray.get(1) is array[1]
PASS webGLArray[0] is 0
PASS webGLArray.get(0) is 0
PASS webGLArray[1] is 0
PASS webGLArray.get(1) is 0
PASS webGLArray[0] is array[0]
PASS webGLArray.get(0) is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray.get(1) is array[1]
Testing WebGLIntArray
PASS webGLArray is array
PASS webGLArray[0] is array[0]
PASS webGLArray.get(0) is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray.get(1) is array[1]
PASS webGLArray[0] is 0
PASS webGLArray.get(0) is 0
PASS webGLArray[1] is 0
PASS webGLArray.get(1) is 0
PASS webGLArray[0] is array[0]
PASS webGLArray.get(0) is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray.get(1) is array[1]
Testing WebGLUnsignedIntArray
PASS webGLArray is array
PASS webGLArray[0] is array[0]
PASS webGLArray.get(0) is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray.get(1) is array[1]
PASS webGLArray[0] is 0
PASS webGLArray.get(0) is 0
PASS webGLArray[1] is 0
PASS webGLArray.get(1) is 0
PASS webGLArray[0] is array[0]
PASS webGLArray.get(0) is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray.get(1) is array[1]
Testing WebGLFloatArray
PASS webGLArray is array
PASS webGLArray[0] is array[0]
PASS webGLArray.get(0) is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray.get(1) is array[1]
PASS webGLArray[0] is 0
PASS webGLArray.get(0) is 0
PASS webGLArray[1] is 0
PASS webGLArray.get(1) is 0
PASS webGLArray[0] is array[0]
PASS webGLArray.get(0) is array[0]
PASS webGLArray[1] is array[1]
PASS webGLArray.get(1) is array[1]
PASS successfullyParsed is true

TEST COMPLETE
